Added more options for driving
I added support for an XBox controller and also for variable speeds
using the triggers on the XBox controller to shift up and down or the
wheel at the bottom of the left joystick in a double-joystick setup.
Also, I added a threshold so that if a joystick is held at 0.9 or higher
it will be held at 1.0 and another to ensure that if two joysticks are
less than 0.1 different in value they will both be read as the higher
number in order to aid in moving straight forward.